+++ src/core/prelude/global.ts
@@ -8,19 +8,23 @@
 
 import log from 'core/log';
 import extend from 'core/prelude/extend';
-import { GLOBAL } from 'core/env';
+import { GLOBAL } from 'core/prelude/env';
+
+if (typeof globalThis === 'undefined') {
+	GLOBAL.globalThis = GLOBAL;
+}
 
 /**
  * Converts the specified unknown value to any
  * @param obj
  */
-extend(GLOBAL, 'Any', (obj) => obj);
+extend(globalThis, 'Any', (obj) => obj);
 
 /**
  * STDERR wrapper
  * @param err
  */
-extend(GLOBAL, 'stderr', (err) => {
+extend(globalThis, 'stderr', (err) => {
 	if (err instanceof Object) {
 		if ({clearAsync: true, abort: true}[String((<Dictionary>err).type)]) {
 			log.info('stderr', err);
@@ -35,4 +39,4 @@ extend(GLOBAL, 'stderr', (err) => {
  * dev/null wrapper
  * @param obj
  */
-extend(GLOBAL, 'devNull', () => { /* empty */ });
+extend(globalThis, 'devNull', () => { /* empty */ });